Added ambulances + fixes

+++ src/event/event.controller.ts
@@ -1,4 +1,5 @@
 import {
+  BadRequestException,
   Body,
   Controller,
   Delete,
@@ -11,6 +12,7 @@ import {
 import { Event as EventModel } from '@prisma/client';
 import { CreateEventDto } from './dto/createEvent.dto';
 import { EventService } from './event.service';
+import { CreateAmbulanceDto } from './dto/createAmbulance.dto';
 
 @Controller('event')
 export class EventController {
@@ -21,8 +23,10 @@ export class EventController {
     @Req() req,
     @Body() createEventDto: CreateEventDto,
   ): Promise<EventModel> {
-    // get tenant ID from the request headers
-    const tenantId = req.headers['tenant-id'];
+    const tenantId = +req.headers['tenant-id'];
+    if (!tenantId || isNaN(tenantId)) {
+      throw new BadRequestException('Tenant ID is invalid');
+    }
     return this.eventService.create({
       ...createEventDto,
       tenant: { connect: { id: tenantId } },
@@ -31,19 +35,28 @@ export class EventController {
 
   @Get()
   async findAll(@Req() req): Promise<EventModel[]> {
-    const tenantId = req.headers['tenant-id'];
+    const tenantId = +req.headers['tenant-id'];
+    if (!tenantId || isNaN(tenantId)) {
+      throw new BadRequestException('Tenant ID is invalid');
+    }
     return this.eventService.findAll(tenantId);
   }
 
   @Get('/:id')
   async findOne(@Req() req, @Param('id') id: number): Promise<EventModel> {
-    const tenantId = req.headers['tenant-id'];
+    const tenantId = +req.headers['tenant-id'];
+    if (!tenantId || isNaN(tenantId)) {
+      throw new BadRequestException('Tenant ID is invalid');
+    }
     return this.eventService.findOne(id, tenantId);
   }
 
   @Delete('/:id')
   async delete(@Req() req, @Param() id: number): Promise<EventModel> {
-    const tenantId = req.headers['tenant-id'];
+    const tenantId = +req.headers['tenant-id'];
+    if (!tenantId || isNaN(tenantId)) {
+      throw new BadRequestException('Tenant ID is invalid');
+    }
     return this.eventService.delete(id, tenantId);
   }
 
@@ -53,7 +66,77 @@ export class EventController {
     @Param('id') id: number,
     @Body() createEventDto: CreateEventDto,
   ): Promise<EventModel> {
-    const tenantId = req.headers['tenant-id'];
+    const tenantId = +req.headers['tenant-id'];
+    if (!tenantId || isNaN(tenantId)) {
+      throw new BadRequestException('Tenant ID is invalid');
+    }
     return this.eventService.update(id, createEventDto, tenantId);
+  }
+
+  @Get('/:id/ambulance')
+  async getAmbulances(@Req() req, @Param('id') id: number) {
+    const tenantId = +req.headers['tenant-id'];
+    if (!tenantId || isNaN(tenantId)) {
+      throw new BadRequestException('Tenant ID is invalid');
+    }
+    return this.eventService.getAmbulances(id, tenantId);
+  }
+
+  @Get('/:id/ambulance/:ambulanceId')
+  async getAmbulance(
+    @Req() req,
+    @Param('id') id: number,
+    @Param('ambulanceId') ambulanceId: number,
+  ) {
+    const tenantId = +req.headers['tenant-id'];
+    if (!tenantId || isNaN(tenantId)) {
+      throw new BadRequestException('Tenant ID is invalid');
+    }
+    return this.eventService.getAmbulance(id, ambulanceId, tenantId);
+  }
+
+  @Post('/:id/ambulance')
+  async createAmbulance(
+    @Req() req,
+    @Param('id') id: number,
+    @Body() createAmbulanceDto: CreateAmbulanceDto,
+  ) {
+    const tenantId = +req.headers['tenant-id'];
+    if (!tenantId || isNaN(tenantId)) {
+      throw new BadRequestException('Tenant ID is invalid');
+    }
+    return this.eventService.createAmbulance(id, createAmbulanceDto, tenantId);
+  }
+
+  @Put('/:id/ambulance/:ambulanceId')
+  async updateAmbulance(
+    @Req() req,
+    @Param('id') id: number,
+    @Param('ambulanceId') ambulanceId: number,
+    @Body() createAmbulanceDto: CreateAmbulanceDto,
+  ) {
+    const tenantId = +req.headers['tenant-id'];
+    if (!tenantId || isNaN(tenantId)) {
+      throw new BadRequestException('Tenant ID is invalid');
+    }
+    return this.eventService.updateAmbulance(
+      id,
+      ambulanceId,
+      createAmbulanceDto,
+      tenantId,
+    );
+  }
+
+  @Delete('/:id/ambulance/:ambulanceId')
+  async deleteAmbulance(
+    @Req() req,
+    @Param('id') id: number,
+    @Param('ambulanceId') ambulanceId: number,
+  ) {
+    const tenantId = +req.headers['tenant-id'];
+    if (!tenantId || isNaN(tenantId)) {
+      throw new BadRequestException('Tenant ID is invalid');
+    }
+    return this.eventService.deleteAmbulance(id, ambulanceId, tenantId);
   }
 }
